SDL2

+++ rts/Game/Game.cpp
@@ -861,7 +861,7 @@ void CGame::ResizeEvent()
 }
 
 
-int CGame::KeyPressed(unsigned short key, bool isRepeat)
+int CGame::KeyPressed(int key, bool isRepeat)
 {
 	if (!gameOver && !isRepeat) {
 		playerHandler->Player(gu->myPlayerNum)->currentStats.keyPresses++;
@@ -926,7 +926,7 @@ int CGame::KeyPressed(unsigned short key, bool isRepeat)
 }
 
 
-int CGame::KeyReleased(unsigned short k)
+int CGame::KeyReleased(int k)
 {
 	if ((userWriting) && (((k>=' ') && (k<='Z')) || (k==8) || (k==190))) {
 		return 0;
