+++ rts/Game/UI/ShareBox.cpp
@@ -14,7 +14,7 @@
 #include "Net/Protocol/NetProtocol.h"
 #include "System/MsgStrings.h"
 
-#include <SDL_keysym.h>
+#include <SDL_keycode.h>
 
 
 #define MAX_SHARE_TEAMS (teamHandler->ActiveTeams() - 1)
@@ -418,7 +418,7 @@ void CShareBox::MouseMove(int x, int y, int dx, int dy, int button)
 	}
 }
 
-bool CShareBox::KeyPressed(unsigned short key, bool isRepeat)
+bool CShareBox::KeyPressed(int key, bool isRepeat)
 {
 	if (key == SDLK_ESCAPE) {
 		delete this;
